# backend/warehouse-service/src/main/resources/static/openapi/warehouse-service.yaml
openapi: 3.0.3
info:
  title: Warehouse Service API
  version: 1.0.0
  description: Contract for warehouse zones, SKU locations, picking tasks, and item movements.
servers:
  - url: http://localhost:2504
paths:
  /api/warehouse/zones:
    post:
      summary: Create a warehouse zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseZoneRequest'
      responses:
        '201':
          description: Zone created
    get:
      summary: Get all warehouse zones
      responses:
        '200':
          description: Zone list
  /api/warehouse/locations:
    post:
      summary: Assign a SKU to a physical location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkuLocationRequest'
      responses:
        '201':
          description: Location assigned
    get:
      summary: Get all SKU locations
      responses:
        '200':
          description: Location list
  /api/warehouse/locations/{sku}:
    get:
      summary: Get warehouse location for a SKU
      parameters:
        - in: path
          name: sku
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Location found
        '404':
          description: Location not found
  /api/warehouse/tasks:
    post:
      summary: Create warehouse picking tasks from an approved order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderTaskRequest'
      responses:
        '201':
          description: Picking tasks created
    get:
      summary: Get all picking tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, IN_PROGRESS, COMPLETED]
      responses:
        '200':
          description: Task list
  /api/warehouse/orders/{orderId}/tasks:
    get:
      summary: Get all tasks for a specific order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order tasks
  /api/warehouse/tasks/{taskId}/start:
    patch:
      summary: Start a picking task
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusUpdateRequest'
      responses:
        '200':
          description: Task started
  /api/warehouse/tasks/{taskId}/complete:
    patch:
      summary: Complete a picking task
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusUpdateRequest'
      responses:
        '200':
          description: Task completed
  /api/warehouse/movements:
    get:
      summary: Get all recorded item movements
      responses:
        '200':
          description: Movement list
  /api/warehouse/movements/sku/{sku}:
    get:
      summary: Get movement history for a SKU
      parameters:
        - in: path
          name: sku
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Movement history
components:
  schemas:
    WarehouseZoneRequest:
      type: object
      required: [code, name, type]
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [STORAGE, RECEIVING, PACKING, STAGING]
        description:
          type: string
    SkuLocationRequest:
      type: object
      required: [sku, zoneCode, shelfCode, onHandQuantity]
      properties:
        sku:
          type: string
        zoneCode:
          type: string
        shelfCode:
          type: string
        onHandQuantity:
          type: integer
          minimum: 0
    OrderTaskRequest:
      type: object
      required: [orderId, shippingAddress, items]
      properties:
        orderId:
          type: integer
          format: int64
        userId:
          type: string
        shippingAddress:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemPayload'
    OrderItemPayload:
      type: object
      required: [sku, quantity]
      properties:
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
    TaskStatusUpdateRequest:
      type: object
      required: [workerId]
      properties:
        workerId:
          type: string